# API_DOCUMENTATION.yaml
openapi: 3.0.0
info:
  title: SIAR Platform ERP/OMS/WMS API
  description: Enterprise-grade Order Management, Inventory Control, and Fulfillment API
  version: 1.0.0
  contact:
    name: SIAR Digital Support
    email: support@siardigital.com
  license:
    name: Proprietary
    url: https://www.siardigital.com
servers:
  - url: http://localhost:5000/api
    description: Development Server
  - url: https://erp.siardigital.com/api
    description: Production Server
...
paths:
...
  /analytics/orders/trends:
    get:
      tags:
        - Analytics
      summary: Order trends analysis
      description: Get daily order counts, revenue, and status breakdown
      parameters:
        - name: days
          in: query
          schema:
            type: integer
            default: 30
        - name: warehouse_id
          in: query
          schema:
            type: integer
      responses:
        '200':
          description: Trends data
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  period_days:
                    type: integer
                  daily_trends:
                    type: array
                    items:
                      type: object
  /analytics/inventory/turnover:
    get:
      tags:
        - Analytics
      summary: Inventory turnover metrics
      description: SKU turnover rates, movement velocity, and stock levels
      parameters:
        - name: days
          in: query
          schema:
            type: integer
            default: 90
      responses:
        '200':
          description: Inventory turnover data
  /analytics/customer/lifetime-value:
    get:
      tags:
        - Analytics
      summary: Customer lifetime value analysis
      description: Customer spending patterns and segmentation
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 100
        - name: min_orders
          in: query
          schema:
            type: integer
            default: 1
      responses:
        '200':
          description: CLV ranking
...
  /analytics/financial/summary:
    get:
      tags:
        - Analytics
      summary: Financial summary
      description: Revenue, COGS, profit, and margins by period
      parameters:
        - name: start_date
          in: query
          schema:
            type: string
            format: date
        - name: end_date
          in: query
          schema:
            type: string
            format: date
      responses:
        '200':
          description: Financial metrics
  /analytics/predictions/demand:
    get:
      tags:
        - Analytics
      summary: Demand forecasting
      description: Predicted demand by SKU for next N days
      parameters:
        - name: days
          in: query
          schema:
            type: integer
            default: 30
        - name: sku_id
          in: query
          schema:
            type: integer
      responses:
        '200':
          description: Demand predictions
  /analytics/reports/custom:
    post:
      tags:
        - Analytics
      summary: Generate custom analytics report
      description: Build custom reports with filters and metrics
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                report_type:
                  type: string
                filters:
                  type: object
                metrics:
                  type: array
      responses:
        '200':
          description: Report generated
  /analytics/export/{reportType}:
    get:
      tags:
        - Analytics
      summary: Export analytics report
      description: Export reports in CSV, JSON, or Excel format
      parameters:
        - name: reportType
          in: path
          required: true
          schema:
            type: string
        - name: format
          in: query
          schema:
            type: string
            enum: [csv, json, excel]
            default: csv
      responses:
        '200':
          description: Report exported

# =====================================================================
# ORDER MANAGEMENT ENDPOINTS (Core - 15+)
# =====================================================================
...
security:
  - BearerToken: []

# =====================================================================
# COMPONENTS & SCHEMAS
# =====================================================================
...